Do not add time or date when in date or time mode
When using 'date' or 'time' mode, the component modifes the data so that the data no longer is only 'date' or 'time'.
When using 'date' mode, initialy the data is for example '2021-05-12', but the component changes the data to '2021-05-12:00:00' or '2021-05-12:00:00:00' (when with-seconds is used). This is not good when working with data in a code that expects '2021-05-12' some problems arise - we need to strip the time again and again.
Similarly in time 'mode'.